requirements.txt
torch
torchvision
onnx
onnxruntime
onnxscript
pillow
numpy
fastapi
pydantic
uvicorn
requests
google-cloud-storage
sqlalchemy